+++ bin/src/pubspec.dart
@@ -1,5 +1,7 @@
 part of '../env_reader.dart';
 
+/// A function to check wether `- assets/env/` already exist in `pubspec.yaml` or not.
+/// If not its inserting the path giver, and if its already there, this will do nothing.
 void insertPubspec() {
   File pubspecFile = File('pubspec.yaml');
   List<String> lines = pubspecFile.readAsLinesSync();
@@ -16,7 +18,8 @@ void insertPubspec() {
   }
 
   if (!foundAssets) {
-    int lastFlutterLineIndex = lines.lastIndexWhere((line) => line.trim() == 'flutter:');
+    int lastFlutterLineIndex =
+        lines.lastIndexWhere((line) => line.trim() == 'flutter:');
     if (lastFlutterLineIndex != -1) {
       lines.insert(lastFlutterLineIndex + 1, '  assets:');
       lines.insert(lastFlutterLineIndex + 2, '    - assets/env/');