+++ lib/env_reader.dart
@@ -6,10 +6,15 @@
 ///
 /// Example:
 /// ```dart
-/// await Env.instance.load();
-/// String? myValue = Env.read<String>("MY_VALUE");
-/// int? myInt = Env.read<int>("MY_INT");
-/// bool? myBool = Env.read<bool>("MY_BOOL");
+/// Future<void> main(List<String> arguments) async {
+///   WidgetsFlutterBinding.ensureInitialized();
+///   await Env.instance.load(password: "my strong password");
+///   runApp(...);
+/// }
+///
+/// String? api = Env.read<String>("API_KEY");
+/// int port = Env.read<int>("PORT") ?? 8080;
+/// bool isDebug = Env.read<bool>("DEBUG") ?? false;
 /// ```
 library env_reader;
 
@@ -27,33 +32,57 @@ class Env {
   ///
   /// Returns the value associated with the specified [key], or `null` if the key
   /// is not found or there was an error while reading the environment.
+  ///
+  /// ```dart
+  /// // example
+  /// final api = Env.read("API_KEY");
+  /// int port = Env.read<int>("PORT") ?? 8080;
+  /// ```
   static T? read<T extends Object>(String key) => instance.read<T>(key);
 }
 
 /// A class for loading and parsing environment variables from a .env file.
 class EnvReader {
-  /// The content of the loaded .env file.
+  /// The raw content of the loaded .env file.
+  ///
+  /// ```dart
+  /// String? env = Env.instance.value;
+  /// ```
   String? value;
 
   /// Loads environment variables from a .env file.
   ///
   /// The [asset] parameter specifies the path to the .env file in asset/env/ directory. By default, it
-  /// sets to look for the assets/env/.env file.
-  Future<void> load({String asset = 'assets/env/.env', required String password}) async {
+  /// sets to look for the `assets/env/.env` file.
+  ///
+  /// ```dart
+  /// Future<void> main(List<String> arguments) async {
+  ///   WidgetsFlutterBinding.ensureInitialized();
+  ///   await Env.instance.load(
+  ///     asset: "assets/env/.env",
+  ///     password: "my strong password");
+  ///   runApp(...);
+  /// }
+  /// ```
+  Future<void> load(
+      {String asset = 'assets/env/.env', required String password}) async {
     try {
       String data = await rootBundle.loadString(asset);
       value = Encryptor.decrypt(password, data);
     } catch (e) {
-      if (kDebugMode) print("\n\n\u001b[1m[ENV_READER]\u001b[31m 💥 Unable to load data\u001b[0m 💥\n$e\n\n");
+      if (kDebugMode) {
+        print(
+            "\n\n\u001b[1m[ENV_READER]\u001b[31m 💥 Unable to load data\u001b[0m 💥\n$e\n\n");
+      }
       value = null;
     }
   }
 
-  /// Parses environment variables into a map.
+  /// Parses environment variables into a json structured map.
   ///
-  /// This method reads the environment variables from the loaded .env file,
-  /// parses their values, and stores them in a map. It returns the map of
-  /// environment variable names and their parsed values.
+  /// ```dart
+  /// Map<String, dynamic> json = Env.instance.toJson();
+  /// ```
   Map<String, dynamic> toJson() {
     final data = <String, dynamic>{};
     final lines = (value ?? "").trim().split('\n');
@@ -80,7 +109,10 @@ class EnvReader {
           }
         }
       } catch (e) {
-        if (kDebugMode) print("\n\n\u001b[1m[ENV_READER]\u001b[31m 💥 Parsing failed\u001b[0m 💥\n$e\n\n");
+        if (kDebugMode) {
+          print(
+              "\n\n\u001b[1m[ENV_READER]\u001b[31m 💥 Parsing failed\u001b[0m 💥\n$e\n\n");
+        }
       }
     }
     return data;
@@ -90,6 +122,13 @@ class EnvReader {
   ///
   /// Returns the value associated with the specified [key], or `null` if the key
   /// is not found or there was an error while reading the environment.
+  ///
+  /// ```dart
+  /// // example
+  /// final env = EnvReader();
+  /// final api = env.read("API_KEY");
+  /// int port = env.read<int>("PORT") ?? 8080;
+  /// ```
   T? read<T extends Object>(String key) {
     try {
       return toJson()[key];